+++ Native2Forms/Notes.iOS/Notes.iOS.csproj
@@ -127,6 +127,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Controllers\AppNavigationController.cs" />
+    <Compile Include="Controllers\NoteDetailsViewController.cs" />
     <Compile Include="Main.cs" />
     <Compile Include="AppDelegate.cs" />
     <Compile Include="Views\NotesPage.xaml.cs">
